+++ src/test/java/stroom/testdata/TestDataGenerator.java
@@ -1,21 +1,32 @@
 package stroom.testdata;
 
-import org.junit.Assert;
-import org.junit.Test;
+import org.assertj.core.api.Assertions;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.io.TempDir;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Path;
 import java.time.Duration;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.Arrays;
 import java.util.Collections;
+import java.util.Objects;
 import java.util.Queue;
 import java.util.concurrent.ConcurrentLinkedQueue;
+import java.util.concurrent.atomic.AtomicInteger;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 public class TestDataGenerator {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(TestDataGenerator.class);
+
+    @TempDir
+    public Path tempDir;
 
     @Test
     public void testCsv_default() {
@@ -31,7 +42,7 @@ public class TestDataGenerator {
                 .generate();
 
         //addition of header row
-        Assert.assertEquals(recCount + 1, lines.size());
+        Assertions.assertThat(lines).hasSize(recCount + 1);
         LOGGER.debug("Dumping generated rows");
         lines.forEach(System.out::println);
     }
@@ -54,9 +65,54 @@ public class TestDataGenerator {
                 .generate();
 
         //no header row so lines == recCount
-        Assert.assertEquals(recCount, lines.size());
+        Assertions.assertThat(lines)
+                .hasSize(recCount);
         LOGGER.debug("Dumping generated rows");
         lines.forEach(System.out::println);
+    }
+
+    @Test
+    public void testMultiLineFileOutputConsumer() {
+        DataGenerator.DefinitionBuilder builder = buildBasicDefinition();
+
+        int recCount = 10;
+        Path outFile = getTempFile();
+        Queue<String> lines = new ConcurrentLinkedQueue<>();
+        builder
+                .setDataWriter(FlatDataWriterBuilder.defaultCsvFormat())
+                .rowCount(recCount)
+                .consumedBy(DataGenerator.getFileOutputConsumer(outFile))
+                .generate();
+
+        //addition of header row
+        Assertions.assertThat(outFile)
+                .isRegularFile();
+
+        final int lineCount = dumpFile(outFile);
+
+        Assertions.assertThat(lineCount).isEqualTo(recCount + 1);
+    }
+
+    @Test
+    public void testSingleLineFileOutputConsumer() {
+        DataGenerator.DefinitionBuilder builder = buildBasicDefinition();
+
+        int recCount = 10;
+        Path outFile = getTempFile();
+        Queue<String> lines = new ConcurrentLinkedQueue<>();
+        builder
+                .setDataWriter(FlatDataWriterBuilder.defaultCsvFormat())
+                .rowCount(recCount)
+                .consumedBy(DataGenerator.getFileOutputConsumer(outFile, "|"))
+                .generate();
+
+        //addition of header row
+        Assertions.assertThat(outFile)
+                .isRegularFile();
+
+        final int lineCount = dumpFile(outFile);
+
+        Assertions.assertThat(lineCount).isEqualTo(1);
     }
 
     @Test
@@ -73,7 +129,8 @@ public class TestDataGenerator {
                 .generate();
 
         //addition of xml declaration, plus opening/closing root elements
-        Assert.assertEquals(recCount + 3, lines.size());
+        Assertions.assertThat(lines)
+                .hasSize(recCount + 3);
         LOGGER.debug("Dumping generated rows");
         lines.forEach(System.out::println);
     }
@@ -96,7 +153,8 @@ public class TestDataGenerator {
                 .generate();
 
         //addition of xml declaration, plus opening/closing root elements
-        Assert.assertEquals(recCount + 3, lines.size());
+        Assertions.assertThat(lines)
+                .hasSize(recCount + 3);
         LOGGER.debug("Dumping generated rows");
         lines.forEach(System.out::println);
     }
@@ -115,7 +173,8 @@ public class TestDataGenerator {
                 .generate();
 
         //addition of xml declaration, plus opening/closing root elements
-        Assert.assertEquals(recCount + 3, lines.size());
+        Assertions.assertThat(lines)
+                .hasSize(recCount + 3);
         LOGGER.debug("Dumping generated rows");
         lines.forEach(System.out::println);
     }
@@ -139,7 +198,8 @@ public class TestDataGenerator {
                 .generate();
 
         //addition of xml declaration, plus opening/closing root elements
-        Assert.assertEquals(recCount + 3, lines.size());
+        Assertions.assertThat(lines)
+                .hasSize(recCount + 3);
         LOGGER.debug("Dumping generated rows");
         lines.forEach(System.out::println);
     }
@@ -157,21 +217,28 @@ public class TestDataGenerator {
         System.out.println("val = " + val);
     }
 
-    @Test (expected = RuntimeException.class)
+    @Test
     public void testRandomWordsField_emptyList() {
-        Field field = DataGenerator.randomWordsField(
-                "myField",
-                2,
-                4,
-                Collections.emptyList());
+        Assertions.assertThatExceptionOfType(RuntimeException.class)
+                .isThrownBy(() -> {
+                    Field field = DataGenerator.randomWordsField(
+                            "myField",
+                            2,
+                            4,
+                            Collections.emptyList());
 
-        String val = field.getNext();
+                    String val = field.getNext();
+                })
+                .withMessageContaining("wordList must have size greater than zero");
     }
 
     private DataGenerator.DefinitionBuilder buildBasicDefinition() {
         //start building a definition that uses all field types
         return DataGenerator.buildDefinition()
                 .multiThreaded()
+                .addFieldDefinition(DataGenerator.fakerField(
+                        "carNumberPlate",
+                        faker -> faker.regexify("[A-Z]{2}[0156][0-9] [A-Z]{3}")))
                 .addFieldDefinition(DataGenerator.sequentialValueField(
                         "sequentialValueField",
                         Arrays.asList("One", "Two", "Three")))
@@ -217,15 +284,42 @@ public class TestDataGenerator {
                         Duration.ofDays(1),
                         "yyyyMMdd"))
                 .addFieldDefinition(DataGenerator.uuidField("uuidField"))
-                .addFieldDefinition(DataGenerator.randomClassNamesField(
-                        "randomClassNamesField",
-                        0,
-                        3))
                 .addFieldDefinition(DataGenerator.randomWordsField(
                         "randomWordsField",
                         0,
                         3,
                         Arrays.asList("attractive", "bald", "beautiful", "chubby", "drab", "elegant", "scruffy", "fit", "glamorous", "handsome", "unkempt")));
+    }
+
+    private Path getTempFile() {
+        try {
+            Path file = Files.createTempFile(tempDir, "stroom-test-", "");
+            LOGGER.info("Using file path {}", file.toAbsolutePath().normalize().toString());
+            return file;
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    private int dumpFile(final Path file) {
+        Objects.requireNonNull(file);
+        Utils.checkArgument(Files.isReadable(file));
+        try {
+            try (Stream<String> stringStream = Files.lines(file)) {
+                final AtomicInteger counter = new AtomicInteger(0);
+                final String content = stringStream
+                        .peek(str -> counter.incrementAndGet())
+                        .collect(Collectors.joining("\n"));
+
+                LOGGER.info("File {} contents:\n{}",
+                        file.toAbsolutePath().normalize().toString(),
+                        content);
+
+                return counter.get();
+            }
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
     }
 
 }