+++ src/test/java/stroom/testdata/TestFlatFileExtension.java
@@ -1,32 +1,28 @@
 package stroom.testdata;
 
-import org.junit.ClassRule;
-import org.junit.Test;
+import org.assertj.core.api.Assertions;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.RegisterExtension;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
 import java.nio.file.Files;
-import java.nio.file.Paths;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.Arrays;
 import java.util.function.Consumer;
 import java.util.stream.Stream;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertTrue;
-import static org.junit.Assert.fail;
-
-public class TestFlatFileRule {
-    private static final Logger LOGGER = LoggerFactory.getLogger(TestFlatFileRule.class);
+public class TestFlatFileExtension {
+    private static final Logger LOGGER = LoggerFactory.getLogger(TestFlatFileExtension.class);
 
     private static final int NUMBER_FILES = 3;
     private static final int ROW_COUNT = 8;
 
-    @ClassRule
-    public static final FlatFileTestDataRule testDataRule = FlatFileTestDataRule.withTempDirectory()
-            .testDataGenerator(TestFlatFileRule::generateTestData)
+    @RegisterExtension()
+    public static final FlatFileTestDataExtension testDataRule = FlatFileTestDataExtension.withTempDirectory()
+            .testDataGenerator(TestFlatFileExtension::generateTestData)
             .numberOfFiles(NUMBER_FILES)
             .build();
 
@@ -51,22 +47,21 @@ public class TestFlatFileRule {
     @Test
     public void testFilesCreated() throws IOException {
         // The rule keeps trak of the files it has created
-        assertEquals(NUMBER_FILES, testDataRule.getDataFiles().size());
+        Assertions.assertThat(testDataRule.getDataFiles().size())
+                .isEqualTo(NUMBER_FILES);
 
         // The rule exposes the folder it has created so you can find the files with a walk
-        Files.walk(Paths.get(testDataRule.getFolder().getAbsolutePath()))
+        Files.walk(testDataRule.getFolder().toAbsolutePath())
                 .filter(p -> !Files.isDirectory(p))
-                .peek(p -> assertTrue(testDataRule.getDataFiles().contains(p)))
+                .peek(p -> Assertions.assertThat(testDataRule.getDataFiles()).contains(p))
                 .forEach(p -> {
                     LOGGER.info("Path Created {}", p);
 
                     try (final Stream<String> stream = Files.lines(p)) {
                         stream.forEach(LOGGER::debug);
                     } catch (IOException e) {
-                        fail(e.getLocalizedMessage());
+                        Assertions.fail(e.getLocalizedMessage());
                     }
                 });
-
-
     }
 }