+++ CHANGELOG.md
@@ -6,9 +6,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
-### Added
+* Add support for writing to file with a custom line separator, e.g. to produce single line data.
 
-### Changed
+* Remove guava.
+
+* Add _java-faker_ value generation.
+
+* Uplift to junit5
 
 * Data Generator consume by function now takes a stream of strings, easier when complex resource based consumption required.